feat(admin): provide an Admin class
We now have an Admin class to apply typographie view to a model admin

This move all the logic to one place.
The fields to be processed need to be declated in the model. It can
handle
strings
dict (for JSONField)
ForeignKey and reverse ForeignKey
OneToOne and reversed OneToOne
ManyToMany and reversed ManyToMany

If you want to apply typographie on a dict you need to provide an
additional list with the list of the keys on which the typographic
filter can be applied.